+++ components/big/tools-sidebar.tsx
@@ -1,5 +1,3 @@
-"use client"
-
 import {
   Sidebar,
   SidebarContent,
@@ -9,6 +7,8 @@ import {
   SidebarMenu,
   SidebarMenuButton,
   SidebarMenuItem,
+  SidebarFooter,
+  SidebarTrigger
 } from "@/components/ui/sidebar"
 import Link from "next/link"
 import { tools } from "@/lib/tools-data"
@@ -39,6 +39,12 @@ export function ToolsSidebar() {
           </SidebarGroupContent>
         </SidebarGroup>
       </SidebarContent>
+      <SidebarFooter className="flex-row md:hidden items-center justify-center">
+        <SidebarTrigger />
+        <div className="font-semibold">
+          Close Sidebar
+        </div>
+      </SidebarFooter>
     </Sidebar>
   )
 }